divide dictionary
